Fix issues with sklearn 1.4
DON'T MERGE YET.

Some skorch tests fail with sklearn v1.4. This commit fixes them:

1. Inheritance structure of scorers seems to have changed.
2. classes_ attribute should always be numpy array
3. CalibratedClassifierCV only works with predict_proba being float64

To fix the latter, I'm now casting the output of predict_proba to
float64. However, I'm not sure if this is a good idea, as it might break
existing code. I'm just adding it in for now to check if the tests pass.

+++ skorch/net.py
@@ -1597,7 +1597,7 @@ class NeuralNet:
             yp = nonlin(yp)
             y_probas.append(to_numpy(yp))
         y_proba = np.concatenate(y_probas, 0)
-        return y_proba
+        return y_proba.astype(np.float64)
 
     def predict(self, X):
         """Where applicable, return class labels for samples in X.
